# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TMS API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "V1"
        ]
      }
    },
...
    "/v2/api/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "V2"
        ]
      }
    },
...
  }
}